combine cache and client.
