+++ client/src/lib.rs
@@ -13,7 +13,6 @@ use thiserror::Error as DeriveError;
 mod actors;
 mod bucket;
 mod client;
-mod external;
 mod ids;
 mod key_store;
 mod provider;
